refactor: styled components를 별개의 파일로 분리
