+++ src/components/common/Chip/Chip.tsx
@@ -1,0 +1,17 @@
+import * as S from "components/common/Chip/Chip.style";
+
+interface ChipProps {
+  children: string;
+  isSelected?: boolean;
+  onClick?: React.MouseEventHandler<HTMLButtonElement>;
+}
+
+function Chip({ children, isSelected, onClick }: ChipProps) {
+  return (
+    <S.ChipContainer isSelected={isSelected} onClick={onClick}>
+      {children}
+    </S.ChipContainer>
+  );
+}
+
+export default Chip;
