--- src/components/common/SearchBar.tsx
@@ -1,60 +1,0 @@
-import React, { useState } from "react";
-import { BsSearch } from "react-icons/bs";
-
-import styled from "styled-components";
-
-const Container = styled.div`
-  display: flex;
-`;
-
-const SearchInput = styled.input`
-  padding: 0.5em 1em;
-
-  background-color: ${({ theme }) => theme.white};
-
-  border: 1px solid ${({ theme }) => theme.secondary};
-  border-radius: 5px;
-  border-top-right-radius: 0;
-  border-bottom-right-radius: 0;
-`;
-
-const Button = styled.button`
-  padding: 0 0.5em;
-
-  background-color: ${({ theme }) => theme.white};
-
-  border: 1px solid ${({ theme }) => theme.secondary};
-  border-radius: 5px;
-  border-top-left-radius: 0;
-  border-bottom-left-radius: 0;
-`;
-
-type SearchBarProps = {
-  onClick: React.MouseEventHandler;
-};
-
-function SearchBar({ onClick }: SearchBarProps) {
-  const [keyword, setKeyword] = useState("");
-
-  const handleSearchInput: React.ChangeEventHandler<HTMLInputElement> = ({
-    target: { value },
-  }) => {
-    setKeyword(value);
-  };
-
-  return (
-    <Container>
-      <SearchInput
-        type="text"
-        placeholder="검색"
-        value={keyword}
-        onChange={handleSearchInput}
-      />
-      <Button onClick={onClick}>
-        <BsSearch />
-      </Button>
-    </Container>
-  );
-}
-
-export default SearchBar;
