+++ src/components/common/Star/Star.style.tsx
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const StarContainer = styled.span`
+  font-size: 1.5rem;
+  color: ${({ theme }) => theme.yellow};
+`;
