+++ src/components/common/StarRating/StarRating.style.tsx
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+export const Container = styled.div`
+  display: flex;
+`;
+
+export const StarButton = styled.button`
+  background-color: transparent;
+  border: none;
+
+  font-size: 1.5rem;
+  color: #e6d706;
+`;
