+++ src/components/common/StoreList/StoreList.style.tsx
@@ -1,0 +1,5 @@
+import styled from "styled-components";
+
+export const StoreListContainer = styled.ul`
+  width: 100%;
+`;
