+++ src/components/layout/Header/Header.tsx
@@ -1,0 +1,46 @@
+import { useReducer } from "react";
+import { BsSearch } from "react-icons/bs";
+import { MdCancel } from "react-icons/md";
+
+import useLogin from "hooks/useLogin";
+
+import SearchBar from "components/common/SearchBar/SearchBar";
+
+import * as S from "components/layout/Header/Header.style";
+
+function Header() {
+  const [isSearchOpen, setSearchOpen] = useReducer(
+    (isOpen: boolean) => !isOpen,
+    false
+  );
+  const [isLogin] = useLogin();
+
+  const handleSearchOpen = () => {
+    setSearchOpen();
+  };
+
+  return (
+    <S.Container isSearchOpen={isSearchOpen}>
+      {isSearchOpen ? (
+        <>
+          <SearchBar onClick={() => {}} />
+          <S.SearchToggleButton onClick={handleSearchOpen}>
+            <MdCancel />
+          </S.SearchToggleButton>
+        </>
+      ) : (
+        <>
+          <S.PageName>mat.zip</S.PageName>
+          <S.RightWrapper>
+            <S.SearchToggleButton onClick={handleSearchOpen}>
+              <BsSearch />
+            </S.SearchToggleButton>
+            {isLogin ? <S.Profile /> : <S.LoginLink>로그인</S.LoginLink>}
+          </S.RightWrapper>
+        </>
+      )}
+    </S.Container>
+  );
+}
+
+export default Header;
