+++ src/components/layout/PageLayout/PageLayout.style.tsx
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export const MainContainer = styled.main`
+  width: 375px;
+  margin: 0 auto;
+
+  background-color: ${({ theme }) => theme.white};
+`;
