+++ src/components/pages/CategoryDetailPage/index.style.tsx
@@ -1,0 +1,15 @@
+import styled from "styled-components";
+
+export const CategoryDetailPageContainer = styled.section`
+  padding: 1em;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  gap: 1em;
+`;
+
+export const ChipContainer = styled.div`
+  display: flex;
+  justify-content: flex-end;
+  gap: 1em;
+`;
