+++ src/components/pages/CategoryPage/Category/Category.style.tsx
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const CategoryContainer = styled.ul`
+  display: grid;
+  grid-template-columns: repeat(4, 1fr);
+  gap: 0.2em;
+  justify-items: center;
+  align-items: center;
+`;
