+++ src/components/pages/CategoryPage/index.style.tsx
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+export const CategoryPageContainer = styled.div`
+  padding: 1em;
+
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  gap: 1.5em;
+`;
