+++ src/components/pages/StoreDetailPage/StoreDetailTitle/StoreDetailTitle.style.tsx
@@ -1,0 +1,21 @@
+import styled from "styled-components";
+
+export const TitleContainer = styled.div`
+  display: flex;
+  width: 100%;
+  flex-direction: column;
+`;
+
+export const TitleRatingWrapper = styled.div`
+  display: flex;
+  width: 100%;
+  flex-direction: row;
+  justify-content: space-between;
+`;
+
+export const DescriptionWrapper = styled.div`
+  width: 100%;
+  overflow: hidden;
+  word-break: break-all;
+  margin-top: 30px;
+`;
