+++ src/stories/Searchbar.stories.jsx
@@ -1,4 +1,4 @@
-import SearchBar from "components/common/SearchBar";
+import SearchBar from "components/common/SearchBar/SearchBar";
 
 export default {
   title: "Components/common/SearchBar",
